+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 
 - Improved CLI output formatting and colorization
+- Default `import` now imports from the `imports` folder in the `data` directory
 
 ### Deprecated